overview.py
import streamlit as st
from PIL import Image

def app():

    st.subheader("Overview")

    st.write('COVID-19, caused by the novel coronavirus SARS-CoV-2, emerged in late 2019 and quickly developed into a global pandemic. It spreads primarily through respiratory droplets when an infected person coughs, sneezes, or talks, and can also be transmitted by touching contaminated surfaces and then touching the face. The virus can lead to a wide range of symptoms, from mild respiratory illness to severe pneumonia and death, with older adults and those with underlying health conditions at higher risk of severe outcomes.')
    st.write('The first cases of COVID-19 in India were reported on 30 January 2020 in three towns of Kerala, among three Indian medical students who had returned from Wuhan, the epicenter of the pandemic. Lockdowns were announced in Kerala on 23 March, and in the rest of the country on 25 March. Infection rates started to drop in September.On the evening of 24 March 2020, the Government of India ordered a nationwide lockdown for 21 days, limiting the movement of the entire 1.38 billion (138 crores) population of India as a preventive measure against the COVID-19 pandemic in India.')
    st.write('In February end 2021, India got hit by the largest COVID wave. It is cited that people started becoming careless, not wearing masks and not following social distancing, around November- April. This wave caused a rapid surge in cases and deaths. Cases started to rise by March 2021, resulting in state-wide lockdowns. In Maharashtra there were total 4 phases of lockdowns from April to June.')

    st.subheader("Preventations")

    st.text('''
        1.Wear Masks 
        2.Practice Social Distancing 
        3.Frequent Hand Hygiene 
        4.Get Vaccinated
        5.Cover your mouth and nose with a bent elbow or tissue when you cough or sneeze. Dispose of used tissues immediately and clean hands regularly  ''')

    img = Image.open('./static/distance.png')
    st.image(img,use_column_width=True)

    st.subheader("Symptoms")

    st.text(''' 1.Fever or chills.
2.Cough.
3.Shortness of breath or difficulty breathing.
4.Fatigue.
5.Muscle or body aches.
6.Headachen.
7.New loss of taste or smell.
8.Sore throat.
9.Congestion or runny nose.
10.Nausea or vomiting.
11.Diarrhea.''')